\name{lamW-package}
\alias{lamW-package}
\alias{lamW}
\docType{package}
\title{
\packageTitle{lamW}
}
\description{
\packageDescription{lamW}
}
\details{
\tabular{ll}{
Package: \tab lamW\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2015-05-19\cr
License: \tab BSD_2_clause \cr
}
}
\author{
Avraham Adler
Maintainer: Avraham Adler <Avraham.Adler@gmail.com>
}
\keyword{package}
\keyword{NumericalMathematics}