\name{lamW-internal}

%% These are the internal C++ functions called by the package
\alias{lambertW0_C}
\alias{lambertWm1_C}
\alias{lamW_RcppExport_registerCCallable}
\alias{lamW_lambertW0_C}
\alias{lamW_lambertWm1_C}

\title{Internal lamW Functions}
\description{
  Internal lambert-W functions
}
\details{
  These are not to be called directly by the user.
}
\keyword{internal}