\name{lamW-package}
\alias{lamW-package}
\alias{lamW}
\docType{package}
\title{
\packageTitle{lamW}
}
\description{
\packageDescription{lamW}
}
\details{
\tabular{ll}{
Package: \tab lamW\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr
Date: \tab 2016-10-26\cr
License: \tab BSD_2_clause \cr
}
}
\author{Avraham Adler \email{Avraham.Adler@gmail.com}}
\keyword{package}
\keyword{NumericalMathematics}