% Copyright (c) 2015, Avraham Adler All rights reserved
% SPDX-License-Identifier: BSD-2-Clause

\name{lamW-package}
\alias{lamW-package}
\alias{lamW}
\docType{package}
\title{
\packageTitle{lamW}
}
\description{
\packageDescription{lamW}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{lamW}
\packageIndices{lamW}
}
\author{
\packageAuthor{lamW}
Maintainer: \packageMaintainer{lamW}
}
\keyword{package}
\keyword{NumericalMathematics}
